Update environment configuration and enhance data quality checks
- Updated .env.example to reflect new PostgreSQL database and user credentials for the data warehouse.
- Modified .gitignore to exclude sensitive configuration files.
- Refactored justfile commands to align with new database settings and improved Docker run configurations.
- Enhanced data quality checks for staging and core layers, including new freshness checks and updated validation logic.
- Adjusted ingestion pipeline to utilize DLT for storing raw data as Parquet on MinIO S3.
- Improved documentation for Airflow DAGs and added event-driven scheduling patterns.

+++ transformations/models/marts/city_weather_metrics.sql
@@ -1,5 +1,6 @@
 -- Mart model: aggregated weather metrics per city.
 -- Provides an overview of each monitored city across all collected data.
+-- Executed in DuckDB, materialized in PostgreSQL via ATTACH.
 
 with current_obs as (
     select
@@ -18,18 +19,18 @@ with current_obs as (
 metrics as (
     select
         city_name,
-        max(country_code)                               as country_code,
-        max(latitude)                                    as latitude,
-        max(longitude)                                   as longitude,
-        count(*)                                         as total_observations,
-        round(avg(temperature_celsius)::numeric, 1)      as avg_temperature_celsius,
-        round(min(temperature_celsius)::numeric, 1)      as min_temperature_celsius,
-        round(max(temperature_celsius)::numeric, 1)      as max_temperature_celsius,
-        round(avg(humidity_percent)::numeric, 1)          as avg_humidity_percent,
-        round(avg(pressure_hpa)::numeric, 1)              as avg_pressure_hpa,
-        round(avg(wind_speed_ms)::numeric, 2)             as avg_wind_speed_ms,
-        min(observed_at)                                  as first_observation,
-        max(observed_at)                                  as last_observation
+        max(country_code)                      as country_code,
+        max(latitude)                           as latitude,
+        max(longitude)                          as longitude,
+        count(*)                                as total_observations,
+        round(avg(temperature_celsius), 1)      as avg_temperature_celsius,
+        round(min(temperature_celsius), 1)      as min_temperature_celsius,
+        round(max(temperature_celsius), 1)      as max_temperature_celsius,
+        round(avg(humidity_percent), 1)          as avg_humidity_percent,
+        round(avg(pressure_hpa), 1)              as avg_pressure_hpa,
+        round(avg(wind_speed_ms), 2)             as avg_wind_speed_ms,
+        min(observed_at)                         as first_observation,
+        max(observed_at)                         as last_observation
     from current_obs
     group by city_name
 )
